+++ src/de.rs
@@ -1,6 +1,6 @@
 use crate::{
-    binary_heap::Kind as BinaryHeapKind, BinaryHeap, Deque, IndexMap, IndexSet, LinearMap, String,
-    Vec,
+    binary_heap::Kind as BinaryHeapKind, BinaryHeap, Deque, HistoryBuffer, IndexMap, IndexSet,
+    LinearMap, String, Vec,
 };
 use core::{
     fmt,
@@ -164,6 +164,43 @@ where
                     if values.push_back(value).is_err() {
                         return Err(A::Error::invalid_length(values.capacity() + 1, &self))?;
                     }
+                }
+
+                Ok(values)
+            }
+        }
+        deserializer.deserialize_seq(ValueVisitor(PhantomData))
+    }
+}
+
+impl<'de, T, const N: usize> Deserialize<'de> for HistoryBuffer<T, N>
+where
+    T: Deserialize<'de>,
+{
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct ValueVisitor<'de, T, const N: usize>(PhantomData<(&'de (), T)>);
+
+        impl<'de, T, const N: usize> serde::de::Visitor<'de> for ValueVisitor<'de, T, N>
+        where
+            T: Deserialize<'de>,
+        {
+            type Value = HistoryBuffer<T, N>;
+
+            fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+                formatter.write_str("a sequence")
+            }
+
+            fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+            where
+                A: SeqAccess<'de>,
+            {
+                let mut values = HistoryBuffer::new();
+
+                while let Some(value) = seq.next_element()? {
+                    values.write(value);
                 }
 
                 Ok(values)