+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added pool implementations for 64-bit architectures.
 - Added `IntoIterator` implementation for `LinearMap`
 - Added `Deque::{get, get_mut, get_unchecked, get_unchecked_mut}`.
+- Added `serde::Serialize` and `serde::Deserialize` implementations to `HistoryBuffer`.
 
 ### Changed